Verify corret test json payload is generated
Will port this to the rest of the tests next. Basically we want to
construct a 'go' version of the request the end user would create.

So we first create a copy of what we would expect to write in go to
create the example request featured on the anki-connect api description.

Then we load the json found on the api description and compare it to
what is generated by the Go structs we have setup.

This actually uncovered an issue where the values of false to bool
fields would be omited. We can resolve this by having *bool in the
structs, but since I want to keep backwards compatability for now, we
just change it so those fields are not omited (and thus default to
false)